Added Status, Changed Help Menu & Winner commands
Added a status for the bot to show what season it's at. The help menu now reflects the new and old features, with addition wording for existing features. Winner functions now support updating winner and loss stats for both regular matches where both players submit memes, and if one does not submit a meme on time