Changed Channel Creation, User submit msg change, winning memes em
Channel creation had a major bug, if the p1 and p2 were in different order, then it made the channel name wrong. Fixed that, User submit msg is sent to mods, so users can't see it anymore. And now there is a winning memes channel, so winning memes get sent there with the User name or nickname